# contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/oncall/rotations": {
      "post": {
        "operationId": "OncallController_createRotation",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create on-call rotation entry",
        "tags": [
          "oncall"
        ]
      },
      "get": {
        "operationId": "OncallController_listRotations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List on-call rotations",
        "tags": [
          "oncall"
        ]
      }
    },
    "/v1/oncall/rotations/{id}": {
      "patch": {
        "operationId": "OncallController_updateRotation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update on-call rotation entry",
        "tags": [
          "oncall"
        ]
      }
    },
    "/v1/oncall/deployments": {
      "post": {
        "operationId": "OncallController_createDeployment",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create on-call deployment entry",
        "tags": [
          "oncall"
        ]
      },
      "get": {
        "operationId": "OncallController_listDeployments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List on-call deployments",
        "tags": [
          "oncall"
        ]
      }
    },
    "/v1/oncall/compliance": {
      "get": {
        "operationId": "OncallController_compliance",
        "parameters": [
          {
            "name": "personId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nextShiftStart",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Evaluate on-call rest compliance",
        "tags": [
          "oncall"
        ]
      }
    },
    "/v1/closing-periods": {
      "get": {
        "operationId": "ClosingController_list",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organizationUnitId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List closing periods",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}": {
      "get": {
        "operationId": "ClosingController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get closing period details",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/start-review": {
      "post": {
        "operationId": "ClosingController_startReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transition closing period from open to review (manual emergency override, admin-only when enabled)",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/lead-approve": {
      "post": {
        "operationId": "ClosingController_leadApprove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record team-lead sign-off for closing period",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/checklist": {
      "get": {
        "operationId": "ClosingController_checklist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate checklist for closing period",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/approve": {
      "post": {
        "operationId": "ClosingController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve closing period",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/export": {
      "post": {
        "operationId": "ClosingController_exportRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "format": {
                    "type": "string",
                    "enum": [
                      "CSV_V1",
                      "XML_V1"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosingExportResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export closing period run with deterministic CSV checksum",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{closingPeriodId}/export-runs/{runId}/csv": {
      "get": {
        "operationId": "ClosingController_downloadCsv",
        "parameters": [
          {
            "name": "closingPeriodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download canonical payroll CSV artifact for an export run",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{closingPeriodId}/export-runs/{runId}/artifact": {
      "get": {
        "operationId": "ClosingController_downloadArtifact",
        "parameters": [
          {
            "name": "closingPeriodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download canonical payroll export artifact for an export run",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/post-close-corrections": {
      "post": {
        "operationId": "ClosingController_postCloseCorrection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create post-close correction workflow",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/corrections/bookings": {
      "post": {
        "operationId": "ClosingController_applyBookingCorrection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workflowId",
                  "personId",
                  "timeTypeId",
                  "startTime",
                  "endTime",
                  "reason"
                ],
                "properties": {
                  "workflowId": {
                    "type": "string"
                  },
                  "personId": {
                    "type": "string"
                  },
                  "timeTypeId": {
                    "type": "string"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Apply approved post-close booking correction (HR/Admin only)",
        "tags": [
          "closing"
        ]
      }
    },
    "/v1/closing-periods/{id}/reopen": {
      "post": {
        "operationId": "ClosingController_reopen",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Re-open closing period from review state (HR only)",
        "tags": [
          "closing"
        ]
      }
    },
...
    "/v1/hr/import-runs/{id}": {
      "get": {
        "operationId": "HrImportController_getImportRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get HR import run by id (integration token required)",
        "tags": [
          "hr-import"
        ]
      }
    },
...
  },
  "info": {
    "title": "cueq API",
    "description": "Integrated time-tracking, absence-management, and shift-planning API for a German university (NRW / TV-L)",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ExportRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "closingPeriodId": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "recordCount": {
            "type": "number"
          },
          "checksum": {
            "type": "string"
          },
          "artifact": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "exportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "exportedById": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "closingPeriodId",
          "format",
          "recordCount",
          "checksum",
          "exportedAt",
          "exportedById"
        ]
      },
      "PayrollCsvRowDto": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          },
          "targetHours": {
            "type": "number"
          },
          "actualHours": {
            "type": "number"
          },
          "balance": {
            "type": "number"
          }
        },
        "required": [
          "personId",
          "targetHours",
          "actualHours",
          "balance"
        ]
      },
      "ClosingExportResponseDto": {
        "type": "object",
        "properties": {
          "exportRun": {
            "$ref": "#/components/schemas/ExportRunDto"
          },
          "checksum": {
            "type": "string"
          },
          "csv": {
            "type": "string",
            "nullable": true
          },
          "artifact": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PayrollCsvRowDto"
            }
          }
        },
        "required": [
          "exportRun",
          "checksum",
          "artifact",
          "contentType",
          "rows"
        ]
      },
...
    }
  }
}